requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==2.0.12
idna==3.4
pyTelegramBotAPI==4.12.0
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2023.3
requests==2.26.0
six==1.16.0
SpeechRecognition==3.10.0
tornado==6.3.2
tzlocal==5.0.1
urllib3==1.26.16
